refactored use of _key_
The TODO suggested use of _key_ on RHS. Because MissingComponent takes a uri as
string as argument in the constructor we use _key_ in order to ensure a correct
type. Then, as a next step we refactored the call to _key_ into a variable and
reused this throughout the function.