remove useless edge case
This is a special case of the check on components[needing].host_uri.
If a service is readonly, then OBVIOUSLY there is a service that depends
on it. If that service is on a host that is "handled", we're going to
need the readonly service. If not, then it's all right to discard it.